actor and critic to encoders
