--- agents/critic.py
@@ -1,17 +1,0 @@
-from abc import abstractmethod
-
-from agents.common import TorchAgent, State, Value, ThreeNodesState
-
-
-class BaseCritic(TorchAgent, config_name='base_critic'):
-    @abstractmethod
-    def forward(self, state: State) -> Value:
-        pass
-
-
-class TowerCritic(BaseCritic, config_name='tower_critic'):
-    def __init__(self):
-        super().__init__()
-
-    def forward(self, state: ThreeNodesState) -> Value:
-        raise NotImplementedError()
